fix(converters): rename MeteoConverter to SpatialConverter and update related tests
